Fix for this:   collectd.c: In function `pidfile_create':   collectd.c:259: warning: int format, pid_t arg (arg 3)

+++ src/collectd.c
@@ -256,7 +256,7 @@ static int pidfile_create (const char *file)
 		return (1);
 	}
 
-	fprintf (fh, "%d\n", getpid());
+	fprintf (fh, "%i\n", (int) getpid ());
 	fclose(fh);
 
 	return (0);
